resource-mapper: use devstack dhcp server IP instead of host vm IP
previously the initial destination IP was the orchestrator host vm IP for
UDP packets generated by the load balancer. Then the resource mapper would
decide where the packets are targeted based on mappings. But if the load
generator is also running in the orchestrator vm, the packets do not even
reach the OVS since they are targeted for the same IP.

The solution is using the devstack's always existing dhcp server ip
(10.0.0.2 by default in our setup) as the initial destination, making sure
all the packets arrive at the OVS and processed with flow rules.