control-panel: inform log-collector after migrating the instance
instance migration causes the target instance to go offline for an
uncertain amount of time, which introduces a delay before updated
flow and group rules can be applied. A simple way for resource-mapper
to detect when the target vm goes online is watching periodical messages
coming from the target vm to log-collector. Upon migration request,
log-collector is informed to watch the target vm and notify resource-
mapper when it's ready